+++ src/main/java/kitchenpos/order/model/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.model;
+package kitchenpos.order.model;
 
 public enum OrderStatus {
     COOKING, MEAL, COMPLETION
