+++ src/test/java/kitchenpos/order/InMemoryOrderLineItemDao.java
@@ -1,7 +1,7 @@
-package kitchenpos.bo;
+package kitchenpos.order;
 
-import kitchenpos.dao.OrderLineItemDao;
-import kitchenpos.model.OrderLineItem;
+import kitchenpos.order.dao.OrderLineItemDao;
+import kitchenpos.order.model.OrderLineItem;
 
 import java.util.*;
 import java.util.stream.Collectors;
