+++ src/test/java/kitchenpos/order/InMemoryOrderTableDao.java
@@ -1,7 +1,7 @@
-package kitchenpos.bo;
+package kitchenpos.order;
 
-import kitchenpos.dao.OrderTableDao;
-import kitchenpos.model.OrderTable;
+import kitchenpos.order.dao.OrderTableDao;
+import kitchenpos.table.model.OrderTable;
 
 import java.util.*;
 import java.util.stream.Collectors;
