+++ src/test/java/kitchenpos/order/controller/OrderRestControllerTest.java
@@ -1,7 +1,7 @@
-package kitchenpos.controller;
+package kitchenpos.order.controller;
 
-import kitchenpos.bo.OrderBo;
-import kitchenpos.model.Order;
+import kitchenpos.order.bo.OrderBo;
+import kitchenpos.order.model.Order;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.web.servlet.HttpEncodingAutoConfiguration;
